# api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: imajin events
  version: "1.0.0"
  description: >
    Event creation, ticketing, ticket tiers, queue management, hold system, .fair manifests, checkout, and payment webhooks.
    Served by the app itself at /api/spec. Authenticate with a scoped app token (`Authorization: Bearer`, audience `events`,
    scopes `events:read` for GET/HEAD/OPTIONS and `events:write` otherwise) or, during migration, the legacy session cookie.
servers:
  - url: https://events.imajin.ai
    description: production
  - url: https://dev-events.imajin.ai
    description: dev
  - url: http://localhost:3000
    description: local
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: imajin_session
    bearerAuth:
      type: http
      scheme: bearer
      description: "Scoped app token minted for audience `events` (events:read for safe methods, events:write otherwise)"
    webhookSecret:
      type: http
      scheme: bearer
      description: "Bearer WEBHOOK_SECRET — internal payment webhook from pay service"
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Not authorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Ticket not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
paths:
...
  /api/webhook/payment:
    post:
      tags: [events]
      operationId: paymentWebhook
      summary: Payment webhook from pay service
      security:
        - webhookSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200":
          description: Event received
...
  /api/orders/{id}/refund:
    post:
      tags: [events]
      operationId: refundOrder
      summary: "Refund an entire order atomically (organizer only)"
      description: "Stripe orders are refunded through the pay service first; free and e-transfer orders are marked directly."
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Order refunded, content: { application/json: { schema: { type: object } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { description: Pay service refund failed; ticket statuses unchanged, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "500": { description: Failed to refund order, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...